creating persistent base template for wishlist and wishlist_item
