+++ service/models/__init__.py
@@ -1,0 +1,8 @@
+"""
+Models for wishlist and wishlist items
+
+All of the models are stored in this package
+"""
+
+from .persistent_base import db, DataValidationError
+from .Wishlist import Wishlist
